+++ R/fonts.R
@@ -427,6 +427,8 @@ font_dep_google_local <- function(x) {
   urls <- extract_group(css, "url\\(([^)]+)")
   basenames <- basename(urls)
 
+  informed <- FALSE
+
   # If need be, download the font file(s) that the CSS imports,
   # and modify the CSS to point to the local files
   Map(function(url, nm) {
@@ -441,7 +443,12 @@ font_dep_google_local <- function(x) {
       x$cache$remove(css_key)
       return(font_dep_google_local(x))
     }
-    download_file(url, f)
+    if (!informed) {
+      rlang::inform(paste0("Downloading google font ", x$family, " to local cache"))
+      informed <<- TRUE
+    }
+    
+    download_file(url, f, quiet = TRUE)
     x$cache$set_file(key, f)
     css <<- sub(url, nm, css, fixed = TRUE)
   }, urls, basenames)
@@ -466,7 +473,8 @@ read_gfont_url <- function(url, file) {
 
   download_file(
     utils::URLencode(url), file,
-    headers = c("User-Agent" = gfont_user_agent())
+    headers = c("User-Agent" = gfont_user_agent()),
+    quiet = TRUE
   )
   readLines(file)
 }
@@ -493,7 +501,8 @@ extract_group <- function(x, pattern, which = 1) {
 # similar to thematic:::download_file, but also translates headers to curl
 #' @importFrom stats na.omit
 #' @importFrom utils download.file packageVersion
-download_file <- function(url, dest, headers = NULL, ...) {
+download_file <- function(url, dest, headers = NULL, quiet = FALSE, ...) {
+
   if (is_installed("curl")) {
     if (!curl::has_internet()) {
       warning(
@@ -505,11 +514,11 @@ download_file <- function(url, dest, headers = NULL, ...) {
       )
     }
     handle <- curl::handle_setheaders(curl::new_handle(), .list = headers)
-    return(curl::curl_download(url, dest, handle = handle, quiet = FALSE, ...))
+    return(curl::curl_download(url, dest, handle = handle, quiet = quiet, ...))
   }
 
   if (capabilities("libcurl")) {
-    return(download.file(url, dest, method = "libcurl", headers = headers, ...))
+    return(download.file(url, dest, method = "libcurl", headers = headers, quiet = quiet, ...))
   }
 
   stop(